+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [27.3.1] - 2025-08-21
+
+### Fixed
+
+- Project
+  - Build scan breaks on zig 0.15+
+  - "Failed to visit file" error blaming ZigBrains when opening projects
+
 ## [27.3.0] - 2025-08-21
 
 ### Added